requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
soupsieve==2.6
urllib3==2.3.0
